First throw, very very much inspired by https://hub.docker.com/r/continuumio/anaconda3
A git clone command has been added, to get a local copy of PyhDToolkit.
The job of creating the appropriate conda environment with `make condaenv` is left to the user for now, as is `pip` installing `cpymad` afterwards.

This will need testing, and later on to be integrated into the build.