fix(kicks): fix the kick implementation
In previous versions of this scripts, when executing for AC dipole kicks, in order to be able to retrieve tunes from tbt data both planes were kicked, one ever so slightly.
However, this influences the second order detuning from sextupolar sources twice as much (see https://doi.org/10.1103/PhysRevSTAB.16.071002) as free oscillations would, this effect has been moved to a small amplitude offset.

Now, AC dipole kick is applied on the given plane while a small offset (J=0.5) is given on the other plane.

Also added in this commit: examples of execution in the module docstring.