refactor: readme & accessories
Refactor README to make it thinner, and include shields because it's what the cool kids do.
The travis file should now report to Code Climate.
Makefile gets rid of docker-pull command which in hindsight was retarded.
Conda environment file removed omc3 which will crash things for now since it is not deployed.